Advance the JMEOS facade to the pin-12l MEOS catalog
Regenerate codegen/input/meos-idl.json from the integration pin (4408
functions) so the binding surface tracks the current MEOS public API:
the base-type operations exposed by MobilityDB#1203, the tjsonb surface,
and the clean standalone base-type names (float8_exp, date_in,
cstring_to_text … replacing the float_/pg_/2 spellings). The build-time
GeneratedFunctions binding is regenerated from this catalog.

Re-derive the MeosOps* facade from the freshly built jar via
regen_facade_from_jar.py, which now resolves each facade method to its
jar-current backing by documented naming convention only (drop the pg_
collision prefix, 2->_to_, float_->float8_) and drops any method with no
backing — no per-function hand map. All thirteen base-type operations
(int32_cmp, int64_cmp, float8_exp, float8_ln, float8_log10,
add_interval_interval, mul_interval_double, minus_date_date,
minus_date_int, minus_timestamptz_interval, minus_timestamptz_timestamptz,
date_to_timestamp, date_to_timestamptz) are retained under their clean
names. Build is green; the test suite is green apart from ten pre-existing
span/box adjacency cases unrelated to this change.

+++ jmeos-core/src/main/java/org/mobilitydb/meos/MeosOpsFreeCbuffer.java
@@ -125,7 +125,7 @@ public final class MeosOpsFreeCbuffer {
      * <p>Pure per-event; safe in any scalar position.</p>
      * <p>Free function (not class-classified). Classification rule: scalar comparison/hash</p>
      */
-    public static long cbuffer_hash_extended(jnr.ffi.Pointer arg0, long arg1) {
+    public static int cbuffer_hash_extended(jnr.ffi.Pointer arg0, int arg1) {
         if (!MEOS_AVAILABLE) {
             throw new UnsupportedOperationException(
                 "cbuffer_hash_extended requires libmeos — set -Dmeos.enabled=true");
