+++ jmeos-core/src/main/java/org/mobilitydb/meos/MeosOpsSet.java
@@ -172,7 +172,7 @@ public final class MeosOpsSet {
      * <p>Object-model role: {@code accessor}.</p>
      * <p>Classification: role=accessor</p>
      */
-    public static long set_hash_extended(jnr.ffi.Pointer arg0, long arg1) {
+    public static int set_hash_extended(jnr.ffi.Pointer arg0, int arg1) {
         if (!MEOS_AVAILABLE) {
             throw new UnsupportedOperationException(
                 "set_hash_extended requires libmeos — set -Dmeos.enabled=true");
